chore: rename InfoBlock to DetailBlock so that InfoBlock can be used for blue messages
