+++ frontend/src/components/common/detail-block/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DetailBlock'
