+++ frontend/src/components/common/index.ts
@@ -7,7 +7,7 @@ export { default as TextInputWithButton } from './text-input-with-button'
 export { default as Pagination } from './pagination'
 export { default as TextArea } from './text-area'
 export { default as FileInput } from './file-input'
-export { default as InfoBlock } from './info-block'
+export { default as DetailBlock } from './detail-block'
 export { default as PrimaryInfoBlock } from './primary-info-block'
 export { default as ErrorBlock } from './error-block'
 export { default as PreviewBlock } from './preview-block'
