--- frontend/src/components/common/info-block/index.ts
@@ -1,1 +1,0 @@
-export { default } from './InfoBlock'
